# src/main/java/de/hsrm/mi/swtp/exchangeplatform/controller/ModulesApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (4.2.1).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package de.hsrm.mi.swtp.exchangeplatform.controller;

import de.hsrm.mi.swtp.exchangeplatform.model.rest_models.Module;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

@javax.annotation.Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2019-12-04T10:25:24.919398+01:00[Europe/Berlin]")

@Validated
@Api(value = "modules", description = "the modules API")
public interface ModulesApi {

    @ApiOperation(value = "", nickname = "getModuleById", notes = "", response = Module.class, tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "single module", response = Module.class) })
    @RequestMapping(value = "/modules/{moduleId}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<Module> getModuleById(@ApiParam(value = "Numeric ID of the module",required=true) @PathVariable("moduleId") Long moduleId);

}